Remove cycles batch I (#7207)
* remove cycles pass 1

* break setupConnection cycle

* break stripe seats dependency loop

* remove lib/data_sources cycle

* break SidebarContenxt import loop

* clean-up

* set to warn

* fixes

* more fixes

* format

* nits
