+++ front/lib/iam/errors.ts
@@ -11,6 +11,7 @@ type AuthFlowErrorCodeType =
   | "invalid_invitation_token"
   | "invitation_token_email_mismatch"
   | "invalid_domain"
+  | "membership_update_error"
   | "revoked";
 
 export class AuthFlowError extends Error {
