+++ front/lib/plans/renderers.ts
@@ -1,0 +1,67 @@
+import type { PlanType, SubscriptionType } from "@dust-tt/types";
+
+import type { Subscription } from "@app/lib/models/plan";
+import type { PlanAttributes } from "@app/lib/plans/free_plans";
+
+// Helper function to render PlanType from PlanAttributes
+export function renderPlanFromModel({
+  plan,
+}: {
+  plan: PlanAttributes;
+}): PlanType {
+  return {
+    code: plan.code,
+    name: plan.name,
+    limits: {
+      assistant: {
+        isSlackBotAllowed: plan.isSlackbotAllowed,
+        maxMessages: plan.maxMessages,
+        maxMessagesTimeframe: plan.maxMessagesTimeframe,
+      },
+      connections: {
+        isConfluenceAllowed: plan.isManagedConfluenceAllowed,
+        isSlackAllowed: plan.isManagedSlackAllowed,
+        isNotionAllowed: plan.isManagedNotionAllowed,
+        isGoogleDriveAllowed: plan.isManagedGoogleDriveAllowed,
+        isGithubAllowed: plan.isManagedGithubAllowed,
+        isIntercomAllowed: plan.isManagedIntercomAllowed,
+        isWebCrawlerAllowed: plan.isManagedWebCrawlerAllowed,
+      },
+      dataSources: {
+        count: plan.maxDataSourcesCount,
+        documents: {
+          count: plan.maxDataSourcesDocumentsCount,
+          sizeMb: plan.maxDataSourcesDocumentsSizeMb,
+        },
+      },
+      users: {
+        maxUsers: plan.maxUsersInWorkspace,
+      },
+      canUseProduct: plan.canUseProduct,
+    },
+    trialPeriodDays: plan.trialPeriodDays,
+  };
+}
+
+// Helper in charge of rendering the SubscriptionType object form PlanAttributes and optionally an
+// active Subscription model.
+export function renderSubscriptionFromModels({
+  plan,
+  activeSubscription,
+}: {
+  plan: PlanAttributes;
+  activeSubscription: Subscription | null;
+}): SubscriptionType {
+  return {
+    status: activeSubscription?.status ?? "active",
+    trialing: activeSubscription?.trialing === true,
+    sId: activeSubscription?.sId || null,
+    stripeSubscriptionId: activeSubscription?.stripeSubscriptionId || null,
+    startDate: activeSubscription?.startDate?.getTime() || null,
+    endDate: activeSubscription?.endDate?.getTime() || null,
+    paymentFailingSince:
+      activeSubscription?.paymentFailingSince?.getTime() || null,
+    plan: renderPlanFromModel({ plan }),
+    requestCancelAt: activeSubscription?.requestCancelAt?.getTime() ?? null,
+  };
+}
