+++ front/lib/plans/usage/seats.ts
@@ -1,9 +1,6 @@
-import type { LightWorkspaceType } from "@dust-tt/types";
 import { cacheWithRedis } from "@dust-tt/types";
-import type Stripe from "stripe";
 
 import { Workspace } from "@app/lib/models/workspace";
-import { updateStripeQuantityForSubscriptionItem } from "@app/lib/plans/stripe";
 import { MembershipResource } from "@app/lib/resources/membership_resource";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 
@@ -32,15 +29,3 @@ export const countActiveSeatsInWorkspaceCached = cacheWithRedis(
   },
   60 * 10 * 1000 // 10 minutes
 );
-
-export async function reportActiveSeats(
-  stripeSubscriptionItem: Stripe.SubscriptionItem,
-  workspace: LightWorkspaceType
-): Promise<void> {
-  const activeSeats = await countActiveSeatsInWorkspace(workspace.sId);
-
-  await updateStripeQuantityForSubscriptionItem(
-    stripeSubscriptionItem,
-    activeSeats
-  );
-}
